Reorganize sdf feature tests
