+++ org.spoofax.jsglr2/src/test/java/org/spoofax/jsglr2/tests/sdf/ListsTest.java
@@ -1,4 +1,4 @@
-package org.spoofax.jsglr2.tests.sdf3;
+package org.spoofax.jsglr2.tests.sdf;
 
 import java.io.IOException;
